# docs/swagger/swagger.yaml
definitions:
...
  AddItemDto:
    properties:
      date_of_manufacture:
        type: string
      medicine_id:
        type: string
      quantity:
        type: number
    required:
    - date_of_manufacture
    - medicine_id
    - quantity
    type: object
...
  AppError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  CabinetItemDetailsResDto:
    properties:
      date_of_manufacture:
        type: string
      expires_at:
        type: string
      id:
        type: string
      medicine_id:
        type: string
      medicine_name:
        type: string
      quantity:
        type: number
    type: object
  CabinetResDto:
    properties:
      id:
        type: string
      medicine_id:
        type: string
      quantity:
        type: number
    type: object
...
  RemoveItemDto:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  SuccessResDTO:
    properties:
      status:
        type: string
    type: object
...
  UpdateQtyDto:
    properties:
      id:
        type: string
      qty:
        type: number
    required:
    - id
    - qty
    type: object
info:
  contact: {}
paths:
...
  /cabinet/items:
    delete:
      consumes:
      - application/json
      description: Удаляет предмет по идентификатору
      parameters:
      - description: Идентификатор предмета
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RemoveItemDto'
      produces:
      - application/json
      responses:
        "200":
          description: Предмет удален
          schema:
            $ref: '#/definitions/SuccessResDTO'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Удаление предмета из кабинета
      tags:
      - Cabinet
    get:
      description: Возвращает предметы кабинета с названиями лекарств и сроками годности
      produces:
      - application/json
      responses:
        "200":
          description: Список предметов
          schema:
            items:
              $ref: '#/definitions/CabinetItemDetailsResDto'
            type: array
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Список предметов кабинета
      tags:
      - Cabinet
    patch:
      consumes:
      - application/json
      description: Изменяет количество предмета в кабинете
      parameters:
      - description: Новое количество
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/UpdateQtyDto'
      produces:
      - application/json
      responses:
        "200":
          description: Количество обновлено
          schema:
            $ref: '#/definitions/CabinetResDto'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Обновление количества
      tags:
      - Cabinet
    post:
      consumes:
      - application/json
      description: Добавляет новый предмет в кабинет пользователя
      parameters:
      - description: Данные предмета
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/AddItemDto'
      produces:
      - application/json
      responses:
        "201":
          description: Предмет добавлен
          schema:
            $ref: '#/definitions/CabinetResDto'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Добавление лекарства в кабинет
      tags:
      - Cabinet
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by your token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"